change py3 code back to py2
I will keep python 3 code in a branch automatically updated by 2to3.

Also removing to load LAS files directly from URLs -- it's easy
enough for users to do that themselves and then provide the text of
the LAS into this package, which is all this aspect of it does anyway
given the way it currently works.