Fix language statistics
GitHub linguistics identified all *.las files as LASSO language and added
them in the statistics of the repo. This small modification in the
.gitattribute corrects this, ignoring the *.las files when calculating the
statistics.

Jupyter Notebooks would still being recognized as the majority of the
files, but I imagine they would vanish as the issues get solved.

TIL: What is
[LASSO](https://en.wikipedia.org/wiki/Lasso_(programming_language))
programming language.