Don't catch Ctrl-C
The general "except:" converted a Ctrl-C into a LASDataError. Therefor
a User abortion of a script using the reader was not possible

General remark: the "except:" is bad and should be replaced by
specific handlers like "except KeyError:" etc. But it is hard to
replace them without the test files that made the original author
insert them