Fix test_encoding_attr, break other tests
This change fixes test_encoding_attr and some other test_encoding.py
tests.

However, it breaks a bunch of other tests, particularly in test_read.py.
Worse, it seems to hang when running test_enhancements.py. It could be
terrible performance or a loop.

So this is just a checkpoint commit, not a resilient solution.