Test for GITHUB_WORKFLOW in verison tests
This change checks to see if Lasio is in a GITHUB_WORKFLOW environment.
If it is then the repo has been cloned without tags and the version
command will return and empty string rather than a git tag version.
If in a GITHUB_WORKFLOW assert the version string is an empty string.