Change remove_line_filter to ignore_data_comments
This commit removes the ability for the user to pass a function
as remove_data_line_filter to LASFile.read, instead replacing
it with the ability to pass:

    - For the header sections, a sequence of strings (`ignore_comments`)
    - For the data sections, a single character (`ignore_data_comments`)

The reason for this change is to align the behaviour
of the normal (current) data section reader with the to-be-implemented
numpy data section reader. The latter will only be capable of
filtering out lines beginning with a single character i.e.
the comments="#" argument to np.genfromtxt.