Reshape data array inside data section reader function
This adds two things:

(1) the argument `n_columns` to the function signature
    for read_data_section_iterative (i.e. the normal data section reader)
    function.

(2) code to (attempt) to re-shape the numpy array read by the
    data section reader into a 2D array.

If the file is wrapped and therefore the number of columns is unclear
(i.e. inspect_data_section() returns -1), then the data section
reader code does not currently attempt to reshape the file.

The reason for this is so that the data section reader function
can be further modified to return a 2D numpy record array with
different dtypes per column.