Allow writing mnemonics in ~A line
Adds mnemonics_header and data_section_header to writer.py:write
The former is boolean controlling whether it occurs or not (off by
default for backwards compatibility). The latter allows the prefix
to be controlled e.g. "~A" vs "~ASCII".

The mnemonic headers are right-aligned based on the numeric value's
column width, so that a fixed-width formatting parser can read it in.
If the mnemonics are too long, then they are just placed one following
the other, out of alignment.