Detect hyphens in data section and adjust regexp_subs as needed
This change modifies defaults.py to add HYPHEN_SUBS - a list of
those substitutions defined elsewhere in this file that look for
and modify matches based on whether there is a hyphen. This is
to discriminate between a substitution which works correctly for
e.g.

123 12.350 1          123 12.350 1
125  6.120 2          125  6.120 2
127 -3.933 3   --->   127 -3.933 3
129-17.000 4          129 -17.000 4
129 -5.001 5          129 -5.001 5

but which falters badly on:

10:15:00 2018-05-22 17.34          10:15:00 2018 -05 -22 17.34
10:15:05 2018-05-22 22.14   --->   10:15:05 2018 -05 -22 22.14
10:15:10 2018-05-22 18.77          10:15:10 2018 -05 -22 18.77

It also modifies reader.py:inspect_data_section such that it

(1) tracks whether a hyphen is found in each line of the data section
    (at least in the inspected part of the data section)
(2) if there is, then it checks whether any of the above substitutions
    (i.e. defaults.py:HYPHEN_SUBS) have been included
(3) if they have been, it returns a list of recommended substitutions
    i.e. all those requested minus those found in HYPHEN_SUBS

Finally las.py:LASFile.read is modified so that:

- it has a new keyword argument accept_regexp_sub_recommendations
  (default True)
- if the recommended substitutions returned by inspect_data_section is different
  to those requested, then if the above keyword argument is True,
  it should re-do the inspection with the recommended changes

As a result, parsing of dates such as '2018-05-22' should be vastly
improved.

A number of docstrings have been unified for usability, also:

- __init__.py:read
- las.py:LASFile.__init__
- las.py:LASFile.read

All tests are passing locally.